refactor: watch コマンドから --stream / --stream-addr フラグを削除する
watch コマンドの --stream / --stream-addr フラグを廃止し、ストリーム配信機能を viewer サブコマンドに完全に集約する。

## 変更内容

- cmd/watch.go: --stream/--stream-addr フラグを hidden に変更し、PreRunE で廃止エラーを返す
- cmd/watch.go: startStreamPlayer 関数と無音モード関連の定数を削除（viewer.go に移動）
- cmd/watch.go: WatchDeps から StreamPlayerFactory フィールドを削除
- cmd/watch.go: runWatch の stream 分岐と silent フラグ関連の処理を削除
- cmd/viewer.go: startStreamPlayer 関数と無音モード定数を追加
- cmd/watch_test.go: ストリーム関連テスト 5 個を削除、廃止フラグ検知テスト 2 個を追加
- cmd/viewer_test.go: stubStreamPlayer と captureStderr ヘルパーを追加
- test/e2e/watch_test.go: ストリーム関連テストケースを廃止フラグテストに更新
- main.go: WatchDeps 初期化から StreamPlayerFactory を削除
- Makefile: dev-backend を `viewer` 呼び出しに変更、run-stream を .PHONY から削除
- docs: watch --stream 言及を viewer に更新（cli.md, README.md, contributing.md）

## テスト結果

- ビルド成功: `go build ./...` OK
- ユニットテスト全てパス: `go test ./...` OK
- Lint 指摘なし: `golangci-lint run` OK

## 受け入れ条件

全て満たしています：
- [x] watch /path でローカルスピーカー再生が動作（既存挙動維持）
- [x] watch --stream は ErrUsage で viewer 案内を表示
- [x] watch --stream-addr も同様に廃止エラー表示
- [x] WatchDeps.StreamPlayerFactory が削除
- [x] ストリーム関連テストが削除・更新
- [x] ドキュメント全て更新
